test: 자동차 일급 컬렉션 테스트
